fix bug with creating memory capabilities
the mark is also of type block_t and should also be used on a page granularity,
this can be observed in cap_mem_derivable where the `tag_block_to_addr` function is used.

This corrects the creation of memory capabilites such that mrk is set to the block value corresponding to bgn, and not the raw address, as previously done.