implement kernel file system driver and file path capability
pulls the full FatFS module into the kernel.

Also adds a crude version of file path capabilities. There is still a lot of work to be done to achieve the proper derivation rules, and also tracking derivations in a way allowing good revocation, a challenge of the current doubly linked list approach.

This commit has a proof of concept in the use-kernel-fs project, which uses kernel apis to read a file's content.