kernel: fix two path revocation bug
previously, the cleared cap table entry was accessed again to retrieve the capability, which would not be correct. it should pass along ccap directly.

secondly, revoking or deleting a capability without children but with siblings did not work correctly. previously it always set the parents first_child equal to the deleted nodes child, which is not correct when there are no children but siblings exist