make bytes_result volatile
it was observed in an example project that results were not consistent with what kernel wrote and what userspace read.

This is either due to weird memory model stuff between kernel and user space, but since it should be executing on the same core I instead assume it is a compiler optimization where it assumes the value does not change. Marking the variable as volatile solves the issue.