add address checking
since we use pointers crossing the user-kernel boundrary the kernel needs to check that it is not writing to or reading from a place the calling process is not allowed to do those operations themselves.

without these, a rogue process could use the new path related capabilities to read from and write to arbitrary addresses using a file storage as an intermediary, or the bytes_result parameter to do more precise modification.