move to an in memory FS for benchmarking purposes
the ORIGIN for apps inside projects needs to be adjusted for the new size of the kernel. Instead of 0x10000 (64 KiB) the kernel takes 0x200000 (2 MiB) of RAM. This means the new BASE for user applications in RAM moves from 0x80010000 to 0x80200000. Projects adapted to this commit will need to search and replace relevant memory addresses.