make creating dir entries benchmark actually exercise space quota checking.
Previously the path had a direct parent to the root, which is tag 0, and the disk usage was not run. This caused the max size quota to perform much better than it should have.

this commit adds an extra directory that all other directories are derviced from, which now does trigger a disk usage check when the directory is created.